requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.7.0
backports.tarfile==1.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
docutils==0.21.2
exceptiongroup==1.2.2
fastapi==0.115.6
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.5
keyring==25.5.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.5.0
nh3==0.2.19
packaging==24.2
pkginfo==1.12.0
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.20
readme_renderer==44.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
sniffio==1.3.1
starlette==0.41.3
twine==6.0.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.34.0
zipp==3.21.0
psycopg2-binary==2.9.9
alembic==1.13.1
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
